# idna maintenance and performance improvements