I started looking at the performance profile of `to_unicode()` and here I am, 12 hours later:

* Updated IDNA tests to latest version
* Updated IDNA mapping table to Unicode 13.0
* Added a `Codec` API that enables some of the allocation costs over multiple `to_unicode()`/`to_ascii()` calls

I wrote some benchmarks for `to_unicode()`, before:

```
test to_unicode_ascii        ... bench:       1,117 ns/iter (+/- 27)
test to_unicode_merged_label ... bench:       4,415 ns/iter (+/- 135)
test to_unicode_puny_label   ... bench:       2,273 ns/iter (+/- 78)
```

and after:

```
test to_unicode_ascii        ... bench:          79 ns/iter (+/- 4)
test to_unicode_merged_label ... bench:       2,034 ns/iter (+/- 58)
test to_unicode_puny_label   ... bench:       1,506 ns/iter (+/- 52)
```

These changes have taken care not to change any of the existing APIs.

Additionally, I saw the old thread about asking for maintenance help. At this point I feel like I'm in a pretty good position to help out with idna maintenance (and could potentially help with the other crates as well if that's useful). Let me know if that'd be useful/how that could take shape.

I've tried to keep commits atomic and easy to review (some of them benefit from reviewing while hiding whitespace changes). Let me know if there are any questions.