@djc 2020-07-09
(Updated original description with benchmark results.)